Adding B-Human compatibility if macro is defined.
B-Human is a C++ code base for RoboCup Standard Platform League soccer,
developed by [B-Human](http://b-human.de) and made available to the
community.  See for more info
[https://github.com/bhuman/BHumanCodeRelease].

My team, the [Dutch Nao Team](http://dutchnaoteam.nl) is using their
framework to develop their own code, and we are using this
DecisionEngine project as a substitute for XABSL.  To keep this engine
open for generic usage, just a macro is added to switch between using
some B-Human code or not.